requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
matplotlib>=3.3.0
Pillow>=8.0.0
PyYAML>=5.4.0
tqdm>=4.62.0
scipy>=1.7.0
einops>=0.4.0
omegaconf>=2.1.0
